requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
wtforms
google-generativeai
werkzeug
pymupdf
email-validator
python-dotenv
